backend/requirements.txt
Flask==3.0.2
Flask-Cors==3.0.10
pymongo==4.6.2
python-dotenv==1.0.1
openai==1.16.2
sentence-transformers==2.2.2